+++ graphics/src/geometry/path/arc.rs
@@ -1,5 +1,5 @@
 //! Build and draw curves.
-use iced_core::{Point, Vector};
+use iced_core::{Point, Radians, Vector};
 
 /// A segment of a differentiable curve.
 #[derive(Debug, Clone, Copy)]
@@ -8,10 +8,10 @@ pub struct Arc {
     pub center: Point,
     /// The radius of the arc.
     pub radius: f32,
-    /// The start of the segment's angle in radians, clockwise rotation from positive x-axis.
-    pub start_angle: f32,
-    /// The end of the segment's angle in radians, clockwise rotation from positive x-axis.
-    pub end_angle: f32,
+    /// The start of the segment's angle, clockwise rotation from positive x-axis.
+    pub start_angle: Radians,
+    /// The end of the segment's angle, clockwise rotation from positive x-axis.
+    pub end_angle: Radians,
 }
 
 /// An elliptical [`Arc`].
@@ -22,11 +22,11 @@ pub struct Elliptical {
     /// The radii of the arc's ellipse. The horizontal and vertical half-dimensions of the ellipse will match the x and y values of the radii vector.
     pub radii: Vector,
     /// The clockwise rotation of the arc's ellipse.
-    pub rotation: f32,
-    /// The start of the segment's angle in radians, clockwise rotation from positive x-axis.
-    pub start_angle: f32,
-    /// The end of the segment's angle in radians, clockwise rotation from positive x-axis.
-    pub end_angle: f32,
+    pub rotation: Radians,
+    /// The start of the segment's angle, clockwise rotation from positive x-axis.
+    pub start_angle: Radians,
+    /// The end of the segment's angle, clockwise rotation from positive x-axis.
+    pub end_angle: Radians,
 }
 
 impl From<Arc> for Elliptical {
@@ -34,7 +34,7 @@ impl From<Arc> for Elliptical {
         Elliptical {
             center: arc.center,
             radii: Vector::new(arc.radius, arc.radius),
-            rotation: 0.0,
+            rotation: Radians(0.0),
             start_angle: arc.start_angle,
             end_angle: arc.end_angle,
         }